requirements.txt
APScheduler==3.6.3
backports.zoneinfo==0.2.1
blis==0.7.5
cachetools==4.2.2
catalogue==2.0.6
certifi==2021.10.8
charset-normalizer==2.0.9
click==8.0.3
cymem==2.0.6
emoji==1.6.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.2.0/en_core_web_sm-3.2.0-py3-none-any.whl
filelock==3.4.0
huggingface-hub==0.2.1
idna==3.3
Jinja2==3.0.3
joblib==1.1.0
langcodes==3.3.0
MarkupSafe==2.0.1
murmurhash==1.0.6
numpy==1.21.4
packaging==21.3
pathy==0.6.1
preshed==3.0.6
protobuf==3.19.1
pydantic==1.8.2
pyparsing==3.0.6
python-telegram-bot==13.8.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2021.11.10
requests==2.26.0
sacremoses==0.0.46
sentencepiece==0.1.96
six==1.16.0
smart-open==5.2.1
spacy==3.2.1
spacy-fi-experimental-web-md==0.8.0
spacy-legacy==3.0.8
spacy-loggers==1.0.1
srsly==2.4.2
stanza==1.3.0
thinc==8.0.13
tokenizers==0.10.3
torch==1.10.0
tornado==6.1
tqdm==4.62.3
transformers==4.12.5
typer==0.4.0
typing_extensions==4.0.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
voikko==0.5
wasabi==0.9.0
